Add filter settings

+++ janitor/clean_files.go
@@ -5,7 +5,7 @@ import (
 	"os"
 )
 
-func CleanFiles(pack string, dir string, keys *Set) {
+func CleanFiles(pack string, dir string, keys *Set, remove bool) {
 	for _, path := range Walk(pack + "/" + dir) {
 		dataPtr, _ := ReadJson[map[string]interface{}](path)
 		fileData := (*dataPtr)
@@ -19,10 +19,12 @@ func CleanFiles(pack string, dir string, keys *Set) {
 				delete(data, key)
 			}
 		}
-
+		if !remove {
+			continue
+		}
 		currentSize := len(data)
 		if currentSize == 0 {
-			fmt.Printf(`Empty file: %s`, path)
+			fmt.Printf("Empty file: %s\n", path)
 			os.Remove(path)
 		} else if currentSize != prevSize {
 			WriteJson(path, dataPtr)
